+++ cooked_input/convertors.py
@@ -15,11 +15,6 @@ from abc import ABCMeta, abstractmethod
 from .error_callbacks import ConvertorError
 
 
-TABLE_ID = 0
-TABLE_VALUE = 1
-TABLE_ID_OR_VALUE = -1
-
-
 ###
 ### Convertors:
 ###